Move isSlotAtNthEpochBoundary() to miscHelpers
