+++ ethereum/spec/src/test/java/tech/pegasys/teku/spec/logic/common/helpers/MiscHelpersTest.java
@@ -22,12 +22,20 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import java.util.stream.Stream;
 import org.apache.tuweni.bytes.Bytes32;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.config.SpecConfig;
+import tech.pegasys.teku.spec.config.SpecConfigLoader;
+import tech.pegasys.teku.spec.networks.Eth2Network;
 
 class MiscHelpersTest {
-  private final SpecConfig specConfig = mock(SpecConfig.class);
+  private final SpecConfig specConfig =
+      SpecConfigLoader.loadConfig(Eth2Network.MINIMAL.configName());
   private final MiscHelpers miscHelpers = new MiscHelpers(specConfig);
 
   @Test
@@ -38,6 +46,9 @@ class MiscHelpersTest {
 
   @Test
   void computeShuffledIndex_samples() {
+    final SpecConfig specConfig = mock(SpecConfig.class);
+    final MiscHelpers miscHelpers = new MiscHelpers(specConfig);
+
     when(specConfig.getShuffleRoundCount()).thenReturn(90);
     assertThat(miscHelpers.computeShuffledIndex(320, 2048, Bytes32.ZERO)).isEqualTo(0);
     assertThat(miscHelpers.computeShuffledIndex(1291, 2048, Bytes32.ZERO)).isEqualTo(1);
@@ -46,6 +57,9 @@ class MiscHelpersTest {
 
   @Test
   void testListShuffleAndShuffledIndexCompatibility() {
+    final SpecConfig specConfig = mock(SpecConfig.class);
+    final MiscHelpers miscHelpers = new MiscHelpers(specConfig);
+
     when(specConfig.getShuffleRoundCount()).thenReturn(10);
     Bytes32 seed = Bytes32.ZERO;
     int index_count = 3333;
@@ -61,6 +75,9 @@ class MiscHelpersTest {
 
   @Test
   void shuffleList_compareListAndArrayVersions() {
+    final SpecConfig specConfig = mock(SpecConfig.class);
+    final MiscHelpers miscHelpers = new MiscHelpers(specConfig);
+
     when(specConfig.getShuffleRoundCount()).thenReturn(10);
     Bytes32 seed = Bytes32.ZERO;
     int index_count = 3333;
@@ -73,5 +90,51 @@ class MiscHelpersTest {
 
     assertThat(result)
         .containsExactlyElementsOf(Arrays.stream(indexes).boxed().collect(Collectors.toList()));
+  }
+
+  @ParameterizedTest(name = "n={0}")
+  @MethodSource("getNValues")
+  void isSlotAtNthEpochBoundary_withSkippedBlock(final int n) {
+    final int nthStartSlot = miscHelpers.computeStartSlotAtEpoch(UInt64.valueOf(n)).intValue();
+
+    final UInt64 genesisSlot = UInt64.ZERO;
+    final UInt64 block1Slot = UInt64.valueOf(nthStartSlot + 1);
+    final UInt64 block2Slot = block1Slot.plus(1);
+    assertThat(miscHelpers.isSlotAtNthEpochBoundary(block1Slot, genesisSlot, n)).isTrue();
+    assertThat(miscHelpers.isSlotAtNthEpochBoundary(block2Slot, block1Slot, n)).isFalse();
+  }
+
+  @ParameterizedTest(name = "n={0}")
+  @MethodSource("getNValues")
+  public void isSlotAtNthEpochBoundary_withSkippedEpochs_oneEpochAndSlotSkipped(final int n) {
+    final int nthStartSlot = miscHelpers.computeStartSlotAtEpoch(UInt64.valueOf(n)).intValue();
+
+    final UInt64 genesisSlot = UInt64.ZERO;
+    final UInt64 block1Slot = UInt64.valueOf(nthStartSlot + specConfig.getSlotsPerEpoch() + 1);
+    final UInt64 block2Slot = block1Slot.plus(1);
+
+    assertThat(miscHelpers.isSlotAtNthEpochBoundary(block1Slot, genesisSlot, n)).isTrue();
+    assertThat(miscHelpers.isSlotAtNthEpochBoundary(block2Slot, block1Slot, n)).isFalse();
+  }
+
+  @ParameterizedTest(name = "n={0}")
+  @MethodSource("getNValues")
+  public void isSlotAtNthEpochBoundary_withSkippedEpochs_nearlyNEpochsSkipped(final int n) {
+    final int startSlotAt2N =
+        miscHelpers.computeStartSlotAtEpoch(UInt64.valueOf(n * 2L)).intValue();
+
+    final UInt64 genesisSlot = UInt64.ZERO;
+    final UInt64 block1Slot = UInt64.valueOf(startSlotAt2N - 1);
+    final UInt64 block2Slot = block1Slot.plus(1);
+    final UInt64 block3Slot = block2Slot.plus(1);
+
+    assertThat(miscHelpers.isSlotAtNthEpochBoundary(block1Slot, genesisSlot, n)).isTrue();
+    assertThat(miscHelpers.isSlotAtNthEpochBoundary(block2Slot, block1Slot, n)).isTrue();
+    assertThat(miscHelpers.isSlotAtNthEpochBoundary(block3Slot, block2Slot, n)).isFalse();
+  }
+
+  public static Stream<Arguments> getNValues() {
+    return Stream.of(
+        Arguments.of(1), Arguments.of(2), Arguments.of(3), Arguments.of(4), Arguments.of(5));
   }
 }
